+++ stable_baselines/common/vec_env/vec_normalize.py
@@ -1,4 +1,5 @@
 import pickle
+import warnings
 
 import numpy as np
 
@@ -9,7 +10,10 @@ from stable_baselines.common.running_mean_std import RunningMeanStd
 class VecNormalize(VecEnvWrapper):
     """
     A moving average, normalizing wrapper for vectorized environment.
-    has support for saving/loading moving average,
+
+    It is pickleable which will save moving averages and configuration parameters.
+    The wrapped environment `venv` is not saved, and must be restored manually with
+    `set_venv` after being unpickled.
 
     :param venv: (VecEnv) the vectorized environment to wrap
     :param training: (bool) Whether to update or not the moving average
@@ -36,6 +40,45 @@ class VecNormalize(VecEnvWrapper):
         self.norm_obs = norm_obs
         self.norm_reward = norm_reward
         self.old_obs = np.array([])
+
+    def __getstate__(self):
+        """
+        Gets state for pickling.
+
+        Excludes self.venv, as in general VecEnv's may not be pickleable."""
+        state = self.__dict__.copy()
+        # these attributes are not pickleable
+        del state['venv']
+        del state['class_attributes']
+        # these attributes depend on the above and so we would prefer not to pickle
+        del state['ret']
+        return state
+
+    def __setstate__(self, state):
+        """
+        Restores pickled state.
+
+        User must call set_venv() after unpickling before using.
+
+        :param state: (dict)"""
+        self.__dict__.update(state)
+        assert 'venv' not in state
+        self.venv = None
+
+    def set_venv(self, venv):
+        """
+        Sets the vector environment to wrap to venv.
+
+        Also sets attributes derived from this such as `num_env`.
+
+        :param venv: (VecEnv)
+        """
+        if self.venv is not None:
+            raise ValueError("Trying to set venv of already initialized VecNormalize wrapper.")
+        VecEnvWrapper.__init__(self, venv)
+        if self.obs_rms.mean.shape != self.observation_space.shape:
+            raise ValueError("venv is incompatible with current statistics.")
+        self.ret = np.zeros(self.num_envs)
 
     def step_wait(self):
         """
@@ -88,10 +131,33 @@ class VecNormalize(VecEnvWrapper):
         self.ret = np.zeros(self.num_envs)
         return self._normalize_observation(obs)
 
+    @staticmethod
+    def load(load_path, venv):
+        """
+        Loads a saved VecNormalize object.
+
+        :param load_path: the path to load from.
+        :param venv: the VecEnv to wrap.
+        :return: (VecNormalize)
+        """
+        with open(load_path, "rb") as file_handler:
+            vec_normalize = pickle.load(file_handler)
+        vec_normalize.set_venv(venv)
+        return vec_normalize
+
+    def save(self, save_path):
+        with open(save_path, "wb") as file_handler:
+            pickle.dump(self, file_handler)
+
     def save_running_average(self, path):
         """
         :param path: (str) path to log dir
+
+        .. deprecated:: 2.9.0
+            This function will be removed in a future version
         """
+        warnings.warn("Usage of `save_running_average` is deprecated. Please "
+                      "use `save` or pickle instead.", DeprecationWarning)
         for rms, name in zip([self.obs_rms, self.ret_rms], ['obs_rms', 'ret_rms']):
             with open("{}/{}.pkl".format(path, name), 'wb') as file_handler:
                 pickle.dump(rms, file_handler)
@@ -99,7 +165,12 @@ class VecNormalize(VecEnvWrapper):
     def load_running_average(self, path):
         """
         :param path: (str) path to log dir
+
+        .. deprecated:: 2.9.0
+            This function will be removed in a future version
         """
+        warnings.warn("Usage of `load_running_average` is deprecated. Please "
+                      "use `load` or pickle instead.", DeprecationWarning)
         for name in ['obs_rms', 'ret_rms']:
             with open("{}/{}.pkl".format(path, name), 'rb') as file_handler:
                 setattr(self, name, pickle.load(file_handler))